refactor(api): move UserNotAuthorizedToUpdateEmailError and InvalidPasswordForUpdateEmailError
to src/shared/application
