+++ api/lib/domain/usecases/update-user-email-with-validation.js
@@ -1,8 +1,5 @@
-import {
-  EmailModificationDemandNotFoundOrExpiredError,
-  InvalidVerificationCodeError,
-  UserNotAuthorizedToUpdateEmailError,
-} from '../errors.js';
+import { UserNotAuthorizedToUpdateEmailError } from '../../../src/shared/domain/errors.js';
+import { EmailModificationDemandNotFoundOrExpiredError, InvalidVerificationCodeError } from '../errors.js';
 
 const updateUserEmailWithValidation = async function ({ code, userId, userEmailRepository, userRepository }) {
   const user = await userRepository.get(userId);
